pyseed/requirements.txt
mysql-connector-python==8.2.0
faker==24.0.0
